requirements2.txt
streamlit
pandas
networkx
matplotlib
python-docx
openpyxl
